requirements.txt
pytest==8.3.3

torch~=2.5.1
PyYAML~=6.0.2
pillow~=11.0.0
torchvision~=0.20.1
matplotlib~=3.9.4
numpy~=1.26.4
ultralytics~=8.3.50
opencv-python~=4.10.0.84
pandas~=2.2.3
seaborn~=0.13.2
scipy~=1.13.1
requests~=2.32.3
tqdm~=4.67.1
psutil~=6.1.0
thop~=0.1.1.post2209072238
ultralytics-thop~=2.0.13